♻️ refactor(auth-files): extract page model helpers
Move auth files search, sort, plan ranking, pagination defaults, and batch animation helpers into a page model module.

Keep AuthFilesPage focused on state, data hooks, and rendering while preserving filtering and sorting behavior.

Validated with type-check, auth files tests, lint, and Vite production build.
